bl/requirement.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
captcha==0.3
certifi==2020.12.5
chardet==4.0.0
chinesecalendar==1.5.0
colorama==0.4.4
cycler==0.10.0
Django==3.1.6
django-filter==2.4.0
django-js-asset==1.2.2
django-matplotlib==0.1
django-mptt==0.11.0
django-pandas==0.6.4
Faker==6.5.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
Jinja2==2.11.3
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mixer==7.1.2
mplfinance==0.12.7a7
my-fake-useragent==0.2.0
mysqlclient==2.0.3
numpy==1.20.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
pluggy==0.13.1
prettytable==2.0.0
py==1.10.0
pyecharts==1.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tushare==1.2.62
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==0.57.0
xlwt==1.3.0
zipp==3.4.0
